Speed up the scraping process.
A lot of time was spent in detecting the character set, parsing HTML
internal routines, and starting new TLS connections.
We now use a C-based character encoding detection library, but in most places
specify which character encoding is used: sometimes we even get it from
the HTTP headers, so we can directly use that. Parsing HTML is solved by
using the lxml library instead of html.parser. We also now use a
connection pool.